requirements.txt
asgiref==3.5.2
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
psycopg2==2.9.5
PyJWT==1.7.1
pytz==2022.6
sqlparse==0.4.3
